backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.3.0
httpx==0.26.0
boto3==1.34.34
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytesseract==0.3.10
google-cloud-vision==3.5.0
openai==1.10.0
pillow==10.2.0
